pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "stacathome"
description = "The WeatherGenerator Dataloader Tool"
readme = "README.md"
version = "0.0.1"
authors = [
    { name = "EarthNet Team" }
]

requires-python = ">=3.11,<3.13"
dependencies = [
    'GDAL==3.11',
    'pdal',
    'geopandas',
    'netCDF4',
    'numpy',
    'odc-geo',
    'odc-stac',
    'pandas',
    'planetary-computer',
    'pyproj',
    'pystac',
    'pystac-client',
    'rasterio',
    'rioxarray',
    'shapely',
    'tqdm',
    'xarray',
    'zarr~=2.18.7',
    'asf_search',
    'rio-stac',
    'pyarrow',
    'earthaccess @ git+https://github.com/MarkusZehner/earthaccess.git@main',
]

[project.optional-dependencies]
dev = [
    'jupyter',
    'ipykernel',
    'dask',
    'jinja2',
    'matplotlib',
]

[dependency-groups]
dev = [
    "pre-commit",
    "pytest",
    "sphinx",
    "coverage",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.black]
skip-string-normalization = true
line-length = 120
target-version = ["py311", "py312", "py313"]


[tool.pytest.ini_options]
testpaths = [
    "test",
]
markers = [
    "remote: marks a test that connects to remote services",
]
